Added some code related to Game Genie functionality. The GUI for entering GG codes isn't complete yet, and the routine for actual patching of code in PSRAM hasn't been written.
Added a tool named stripcom that strips comments from intermediate assembly files. The reason for this is that the opt-816.py script was having trouble reading the assembly files because the snes C compiler was putting comments with strange characters in them.
C code for the tool is included (it's very short).

Modified the joypad mapping a bit. Left/right (D-pad) now steps one page in the games list, just like in the MD menu. L (shoulder button) is used to toggle alphabetical listing. Y is now generally used as the "back" button, and B as the "action" button.