Game Genie functionality is now working.
Replaced the font.

Modified the controls a bit.

Beware of the space limitations. There's about 16kB of room left right now. If more space is needed it will have to come in the form of more ROM-banks, and then the snes-sdk libs will also have to be recompiled with the new header.